warsan/requirements.txt
appdirs==1.4.4
arrow==1.2.3
asgiref==3.7.2
attrs==23.1.0
black==21.5b1
click==8.1.7
Django==3.2.6
django-filter==23.2
django-phonenumber-field @ git+https://github.com/stefanfoulis/django-phonenumber-field.git@43958a39e0b48ad6e2a4b86c0ee06bdb175d638d
djangorestframework==3.14.0
flake8==3.9.2
gunicorn==20.1.0
iniconfig==2.0.0
Jinja2==3.1.2
jinja2-time==0.2.0
make==0.1.6.post2
Markdown==3.4.4
MarkupSafe==2.1.3
mccabe==0.6.1
migrations==0.1.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
phonenumbers==8.13.20
phonenumberslite==8.13.20
pluggy==0.13.1
psycopg2==2.9.1
py==1.11.0
pycodestyle==2.7.0
pyflakes==2.3.1
pytest==6.2.4
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.8.8
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.7.1
